Create edit_distance.java
Edit distance is a term used in computer science. By definition, Edit distance is a string metric, a way of quantifying how dissimilar two strings (e.g. words) are to one another, measured by counting the minimum number of operations required to transform one string into the other. Source
What also makes it more interesting to solve this problem is that this problem has several real-world applications. Think of it like this, if you want to find the similarity score between 2 strings. The lower the edit distance, the more similar they are.